feat(cli): new subcommands (#283)
### Problem Statement

We may want to use parts of this build system with parts of our own build system. I may have use of using my own build script to call rollup and tsc.

### Solution

To start this conversation I suggest we make new subcommands for validating and doing other things step-by-step. This is somewhat designed as a sketch to gather feedback and see if you are interested in this kind of work.

### How has this been tested?

starting to manually run some of the tests that run in CI noticed a bug

something going on with CI build was not working for some reason
---

[related issue](https://github.com/near/near-sdk-js/issues/280)

Co-authored-by: Serhii Volovyk <SergeyVolovyk@gmail.com>