Remove alpha warning
Since we have done testing regarding arithmetic correctness in https://github.com/near/libbf-test, in the same test suite for the C++'s boost.multiprecision library. We consider the JavaScript runtime is reasonably secure for DApps and even financial DApps. The alpha warning is therefore removed. 

The contributing parts at beginning of the doc is duplicate to the `## Contributing` section later, so it's also removed