Implement improvements as suggested in the last report
🚨Attention: before merging, this patch needs to be squashed with a
future commit which changes where we fetch wasi-sdk binaries, now we
need to fetch from a near fork of wasi-sdk.

This patch includes improvements as suggested in the last report. Most
notably, it drops a very legacy version of wasi-sdk and is able to
make use of latest wasi-sdk. The wasi-sdk patch makes things work out
of the box for the near platform, allowing the wasi-stub hack to be
dropped here as well, so no need to maintain a binaryen fork either.